requirements.txt
flask
psutil
requests
dnspython
speedtest-cli
scapy
pandas         # for CSV export
reportlab      # for generating PDFs (or use WeasyPrint or FPDF instead)
